Added joins for ClientManager Threads

+++ libs/sub_server/ActionCreateServer.cpp
@@ -6,7 +6,7 @@
 #include <utility>
 #include "sub_common/protocolo.h"
 
-ActionCreateServer::ActionCreateServer(uint8_t &id, int &c, const std::string &name, GameManager &gameManager,
+ActionCreateServer::ActionCreateServer(uint8_t &id, uint8_t &c, const std::string &name, GameManager &gameManager,
                                        std::function<void(BlockingQueue<Action *> *, BlockingQueue<Action *> *)> function) :
                            Action(id), gameManager(gameManager),startClientThreads(std::move(function)),capacityGame(c), nameGame(name) {}
 
@@ -14,7 +14,7 @@ std::string ActionCreateServer::getGameName() {
     return nameGame;
 }
 
-int ActionCreateServer::getCapacity() const {
+uint8_t ActionCreateServer::getCapacity() const {
     return capacityGame;
 }
